+++ src/main/java/ui/pages/DepositMoneyPage.java
@@ -4,14 +4,16 @@ import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.Selectors;
 import com.codeborne.selenide.SelenideElement;
 
+import java.math.BigDecimal;
+
 import static com.codeborne.selenide.Selenide.$;
 
 public class DepositMoneyPage extends BasePage<DepositMoneyPage> {
     private final SelenideElement selectAccountField = $(".account-selector");
     private final SelenideElement amountField = $(".deposit-input");
     private final SelenideElement depositButton = $(Selectors.byText("\uD83D\uDCB5 Deposit"));
-    public final static float DEFAULT_BALANCE = 0.00f;
-    public final static float NEGATIVE_MAXIMUM_BOUNDARY_VALUE = 5000.01f;
+    public final static BigDecimal DEFAULT_BALANCE = BigDecimal.valueOf(0.00);
+    public final static BigDecimal NEGATIVE_MAXIMUM_BOUNDARY_VALUE = BigDecimal.valueOf(5000.01);
 
     @Override
     public String url() {
@@ -26,7 +28,7 @@ public class DepositMoneyPage extends BasePage<DepositMoneyPage> {
         return this;
     }
 
-    public DepositMoneyPage enterDepositAmount(float depositAmount) {
+    public DepositMoneyPage enterDepositAmount(BigDecimal depositAmount) {
         amountField.sendKeys(String.valueOf(depositAmount));
         return this;
     }
@@ -36,7 +38,7 @@ public class DepositMoneyPage extends BasePage<DepositMoneyPage> {
         return new UserDashboard();
     }
 
-    public DepositMoneyPage checkAccountBalance(Long accountNumber, float depositAmount) {
+    public DepositMoneyPage checkAccountBalance(Long accountNumber, BigDecimal depositAmount) {
         selectDepositAccount(accountNumber);
         selectAccountField.shouldHave(Condition.text(String.valueOf(depositAmount)));
         return this;
